tests: Move integration tests to their own directory
Simplify main.rs by moving the integration tests to their own directory.
